Remove page-size from Pagination and HTTP headers.
1) This is not needed because it is derivable from 2 other properties:
first-record and last-record: last - first + 1;
2) This will lead to bugs if used on the client side incorrectly:
page-size might be interpreted as an expected size for all pages; but it
is calculated based on actual size. Thus, when one navigates to the
last page, which may have less items to display, the page size would be
updated to this new, smaller value - altering the entire pagination
calculation. For example:
- total items: 12
- (all pages) page size: 10
On page 1, the number of records is 10. Page size is calculated as 10.
Clicking 'next' will get one to page 2: records 11-12.
Now, on page 2, the number of records is 2. Page size is calculated as
2. Click 'previous' will get one to page 1. But since page size has been
set to 2, page 1 now is displaying records 9-10. Depending on how the
page number is set, the page number may change too.